+++ sentences/HDG.js
@@ -8,6 +8,7 @@ $IIHDG,x.x,,,,*hh
 const nmea = require('../nmea.js')
 module.exports = function (app) {
   return {
+    sentence: 'HDG',
     title: 'HDG - Heading magnetic:.',
     keys: ['navigation.headingMagnetic', 'navigation.magneticVariation' ],
     defaults: [undefined, ''],