+++ sentences/MWVR.js
@@ -21,12 +21,13 @@
 const nmea = require('../nmea.js')
 module.exports = function (app) {
   return {
+    sentence: 'MWV',
     title: 'MWV - Aparent Wind heading and speed',
     keys: ['environment.wind.angleApparent', 'environment.wind.speedApparent'],
     f: function (angle, speed) {
       return nmea.toSentence([
-        '$INMWV',
-        nmea.radsToDeg(angle).toFixed(2),
+        '$IIMWV',
+        nmea.radsToPositiveDeg(angle).toFixed(2),
         'R',
         speed.toFixed(2),
         'M',