+++ sentences/MWVT.js
@@ -3,13 +3,14 @@
 const nmea = require('../nmea.js')
 module.exports = function (app) {
   return {
+    sentence: 'MWV',
     title: 'MWV - True Wind heading and speed',
-    keys: ['environment.wind.angleTrue', 'environment.wind.speedTrue'],
+    keys: ['environment.wind.angleTrueWater', 'environment.wind.speedTrue'],
 
     f: function (angle, speed) {
       return nmea.toSentence([
-        '$INMWV',
-        nmea.radsToDeg(angle).toFixed(2),
+        '$IIMWV',
+        nmea.radsToPositiveDeg(angle).toFixed(2),
         'T',
         speed.toFixed(2),
         'M',