+++ test/RMC.js
@@ -1,13 +1,14 @@
-const Bacon = require('baconjs')
 const assert = require('assert')
+
+const {createAppWithPlugin} = require ('./testutil')
 
 describe('RMC', function () {
   it('works without datetime & magneticVariation', done => {
     const onEmit = (event, value) => {
-      assert.equal(value, '$SKRMC,,A,0600.0000,N,00500.0000,E,1.9,114.6,,,E*5E')
+      assert.equal(value, '$GPRMC,,A,0600.0000,N,00500.0000,E,1.9,114.6,,,E*51')
       done()
     }
-    const app = createAppWithPlugin(onEmit)
+    const app = createAppWithPlugin(onEmit, 'RMC')
     app.streambundle.getSelfStream('navigation.speedOverGround').push('1')
     app.streambundle.getSelfStream('navigation.courseOverGroundTrue').push('2')
     app.streambundle
@@ -15,14 +16,15 @@ describe('RMC', function () {
       .push({ longitude: 5, latitude: 6 })
   })
 
-  it('works with large longitude', done => {
+  it('works with large longitude & magnetic variation', done => {
     const onEmit = (event, value) => {
-      assert.equal(value, '$SKRMC,,A,3749.6038,N,12225.2480,W,1.9,114.6,,,E*43')
+      assert.equal(value, '$GPRMC,,A,3749.6038,N,12225.2480,W,1.9,114.6,,180.0,E*6B')
       done()
     }
-    const app = createAppWithPlugin(onEmit)
+    const app = createAppWithPlugin(onEmit, 'RMC')
     app.streambundle.getSelfStream('navigation.speedOverGround').push('1')
     app.streambundle.getSelfStream('navigation.courseOverGroundTrue').push('2')
+    app.streambundle.getSelfStream('navigation.magneticVariation').push(Math.PI)
     app.streambundle
       .getSelfStream('navigation.position')
       .push({ longitude: -122.4208, latitude: 37.82673 })
@@ -30,10 +32,10 @@ describe('RMC', function () {
 
   it('ignores a too large longitude', done => {
     const onEmit = (event, value) => {
-      assert.equal(value, '$SKRMC,,A,3749.6038,N,12225.2480,W,1.9,114.6,,,E*43')
+      assert.equal(value, '$GPRMC,,A,3749.6038,N,12225.2480,W,1.9,114.6,,,E*4C')
       done()
     }
-    const app = createAppWithPlugin(onEmit)
+    const app = createAppWithPlugin(onEmit, 'RMC')
     app.streambundle.getSelfStream('navigation.speedOverGround').push('1')
     app.streambundle.getSelfStream('navigation.courseOverGroundTrue').push('2')
     app.streambundle
@@ -47,23 +49,3 @@ describe('RMC', function () {
     }, 50)
   })
 })
-
-function createAppWithPlugin (onEmit) {
-  const streams = {
-    'navigation.speedOverGround': new Bacon.Bus(),
-    'navigation.courseOverGroundTrue': new Bacon.Bus(),
-    'navigation.datetime': new Bacon.Bus(),
-    'navigation.position': new Bacon.Bus(),
-    'navigation.magneticVariation': new Bacon.Bus()
-  }
-  const app = {
-    streambundle: { getSelfStream: path => streams[path] },
-    emit: onEmit
-  }
-  const plugin = require('../')(app)
-  const options = {
-    RMC: true
-  }
-  plugin.start(options)
-  return app
-}